separate tasks tests and simplify tests
