+++ common/persistence/data_manager_interfaces_test.go
@@ -78,128 +78,6 @@ func TestIsBackgroundTransientError(t *testing.T) {
 	}
 }
 
-func TestTaskCommonMethods(t *testing.T) {
-	timeNow := time.Now()
-	tasks := []Task{
-		&ActivityTask{TaskData: TaskData{Version: 1, TaskID: 1, VisibilityTimestamp: timeNow}},
-		&DecisionTask{TaskData: TaskData{Version: 1, TaskID: 1, VisibilityTimestamp: timeNow}},
-		&RecordWorkflowStartedTask{TaskData: TaskData{Version: 1, TaskID: 1, VisibilityTimestamp: timeNow}},
-		&ResetWorkflowTask{TaskData: TaskData{Version: 1, TaskID: 1, VisibilityTimestamp: timeNow}},
-		&CloseExecutionTask{TaskData: TaskData{Version: 1, TaskID: 1, VisibilityTimestamp: timeNow}},
-		&DeleteHistoryEventTask{TaskData: TaskData{Version: 1, TaskID: 1, VisibilityTimestamp: timeNow}},
-		&DecisionTimeoutTask{TaskData: TaskData{Version: 1, TaskID: 1, VisibilityTimestamp: timeNow}},
-		&ActivityTimeoutTask{TaskData: TaskData{Version: 1, TaskID: 1, VisibilityTimestamp: timeNow}},
-		&UserTimerTask{TaskData: TaskData{Version: 1, TaskID: 1, VisibilityTimestamp: timeNow}},
-		&ActivityRetryTimerTask{TaskData: TaskData{Version: 1, TaskID: 1, VisibilityTimestamp: timeNow}},
-		&WorkflowBackoffTimerTask{TaskData: TaskData{Version: 1, TaskID: 1, VisibilityTimestamp: timeNow}},
-		&WorkflowTimeoutTask{TaskData: TaskData{Version: 1, TaskID: 1, VisibilityTimestamp: timeNow}},
-		&CancelExecutionTask{TaskData: TaskData{Version: 1, TaskID: 1, VisibilityTimestamp: timeNow}},
-		&SignalExecutionTask{TaskData: TaskData{Version: 1, TaskID: 1, VisibilityTimestamp: timeNow}},
-		&RecordChildExecutionCompletedTask{TaskData: TaskData{Version: 1, TaskID: 1, VisibilityTimestamp: timeNow}},
-		&ApplyParentClosePolicyTask{TaskData: TaskData{Version: 1, TaskID: 1, VisibilityTimestamp: timeNow}},
-		&UpsertWorkflowSearchAttributesTask{TaskData: TaskData{Version: 1, TaskID: 1, VisibilityTimestamp: timeNow}},
-		&StartChildExecutionTask{TaskData: TaskData{Version: 1, TaskID: 1, VisibilityTimestamp: timeNow}},
-		&RecordWorkflowClosedTask{TaskData: TaskData{Version: 1, TaskID: 1, VisibilityTimestamp: timeNow}},
-		&HistoryReplicationTask{TaskData: TaskData{Version: 1, TaskID: 1, VisibilityTimestamp: timeNow}},
-		&SyncActivityTask{TaskData: TaskData{Version: 1, TaskID: 1, VisibilityTimestamp: timeNow}},
-		&FailoverMarkerTask{TaskData: TaskData{Version: 1, TaskID: 1, VisibilityTimestamp: timeNow}},
-	}
-
-	for _, task := range tasks {
-		switch ty := task.(type) {
-		case *ActivityTask:
-			assert.Equal(t, TransferTaskTypeActivityTask, ty.GetType())
-		case *DecisionTask:
-			assert.Equal(t, TransferTaskTypeDecisionTask, ty.GetType())
-		case *RecordWorkflowStartedTask:
-			assert.Equal(t, TransferTaskTypeRecordWorkflowStarted, ty.GetType())
-		case *ResetWorkflowTask:
-			assert.Equal(t, TransferTaskTypeResetWorkflow, ty.GetType())
-		case *CloseExecutionTask:
-			assert.Equal(t, TransferTaskTypeCloseExecution, ty.GetType())
-		case *DeleteHistoryEventTask:
-			assert.Equal(t, TaskTypeDeleteHistoryEvent, ty.GetType())
-		case *DecisionTimeoutTask:
-			assert.Equal(t, TaskTypeDecisionTimeout, ty.GetType())
-		case *ActivityTimeoutTask:
-			assert.Equal(t, TaskTypeActivityTimeout, ty.GetType())
-		case *UserTimerTask:
-			assert.Equal(t, TaskTypeUserTimer, ty.GetType())
-		case *ActivityRetryTimerTask:
-			assert.Equal(t, TaskTypeActivityRetryTimer, ty.GetType())
-		case *WorkflowBackoffTimerTask:
-			assert.Equal(t, TaskTypeWorkflowBackoffTimer, ty.GetType())
-		case *WorkflowTimeoutTask:
-			assert.Equal(t, TaskTypeWorkflowTimeout, ty.GetType())
-		case *CancelExecutionTask:
-			assert.Equal(t, TransferTaskTypeCancelExecution, ty.GetType())
-		case *SignalExecutionTask:
-			assert.Equal(t, TransferTaskTypeSignalExecution, ty.GetType())
-		case *RecordChildExecutionCompletedTask:
-			assert.Equal(t, TransferTaskTypeRecordChildExecutionCompleted, ty.GetType())
-		case *ApplyParentClosePolicyTask:
-			assert.Equal(t, TransferTaskTypeApplyParentClosePolicy, ty.GetType())
-		case *UpsertWorkflowSearchAttributesTask:
-			assert.Equal(t, TransferTaskTypeUpsertWorkflowSearchAttributes, ty.GetType())
-		case *StartChildExecutionTask:
-			assert.Equal(t, TransferTaskTypeStartChildExecution, ty.GetType())
-		case *RecordWorkflowClosedTask:
-			assert.Equal(t, TransferTaskTypeRecordWorkflowClosed, ty.GetType())
-		case *HistoryReplicationTask:
-			assert.Equal(t, ReplicationTaskTypeHistory, ty.GetType())
-		case *SyncActivityTask:
-			assert.Equal(t, ReplicationTaskTypeSyncActivity, ty.GetType())
-		case *FailoverMarkerTask:
-			assert.Equal(t, ReplicationTaskTypeFailoverMarker, ty.GetType())
-		default:
-			t.Fatalf("Unhandled task type: %T", t)
-		}
-
-		// Test version methods
-		assert.Equal(t, int64(1), task.GetVersion())
-		task.SetVersion(2)
-		assert.Equal(t, int64(2), task.GetVersion())
-
-		// Test TaskID methods
-		assert.Equal(t, int64(1), task.GetTaskID())
-		task.SetTaskID(2)
-		assert.Equal(t, int64(2), task.GetTaskID())
-
-		// Test VisibilityTimestamp methods
-		assert.Equal(t, timeNow, task.GetVisibilityTimestamp())
-		newTime := timeNow.Add(time.Second)
-		task.SetVisibilityTimestamp(newTime)
-		assert.Equal(t, newTime, task.GetVisibilityTimestamp())
-	}
-}
-
-func TestOnlyGetTypeTask(t *testing.T) {
-	tasks := []Task{
-		&CrossClusterStartChildExecutionTask{},
-		&CrossClusterCancelExecutionTask{},
-		&CrossClusterSignalExecutionTask{},
-		&CrossClusterRecordChildExecutionCompletedTask{},
-		&CrossClusterApplyParentClosePolicyTask{},
-	}
-
-	for _, task := range tasks {
-		switch ty := task.(type) {
-		case *CrossClusterStartChildExecutionTask:
-			assert.Equal(t, CrossClusterTaskTypeStartChildExecution, ty.GetType())
-		case *CrossClusterCancelExecutionTask:
-			assert.Equal(t, CrossClusterTaskTypeCancelExecution, ty.GetType())
-		case *CrossClusterSignalExecutionTask:
-			assert.Equal(t, CrossClusterTaskTypeSignalExecution, ty.GetType())
-		case *CrossClusterRecordChildExecutionCompletedTask:
-			assert.Equal(t, CrossClusterTaskTypeRecordChildExeuctionCompleted, ty.GetType())
-		case *CrossClusterApplyParentClosePolicyTask:
-			assert.Equal(t, CrossClusterTaskTypeApplyParentClosePolicy, ty.GetType())
-		default:
-			t.Fatalf("Unhandled task type: %T", t)
-		}
-	}
-}
-
 func TestNewHistoryBranch(t *testing.T) {
 	newTreeID := "newTreeID"
 	generalToken, err := NewHistoryBranchToken(newTreeID)
